[NET-765] Global database module renaming
