+++ cmdb/__main__.py
@@ -34,7 +34,7 @@ except ImportError:
     CMDBError = Exception
 
 from cmdb.utils.system_config import SystemConfigReader
-from cmdb.data_storage.database_manager import DatabaseManagerMongo
+from cmdb.database.database_manager import DatabaseManagerMongo
 
 # setup logging for startup
 logging.config.dictConfig(get_logging_conf())
@@ -56,7 +56,7 @@ def _check_database():
     Returns: Datebase if response
 
     """
-    from cmdb.data_storage.errors.connection_errors import ServerTimeoutError
+    from cmdb.database.errors.connection_errors import ServerTimeoutError
     ssc = SystemConfigReader()
     LOGGER.info(f'Checking database connection with {ssc.config_name} data')
     database_options = ssc.get_all_values_from_section('Database')
@@ -154,7 +154,7 @@ def main(args):
     if args.debug:
         _activate_debug()
     _init_config_reader(args.config_file)
-    from cmdb.data_storage.errors.connection_errors import DatabaseConnectionError
+    from cmdb.database.errors.connection_errors import DatabaseConnectionError
 
     # create / check connection database managers
     dbm = None
