+++ cmdb/database/errors/__init__.py
