+++ cmdb/database/errors/connection_errors.py
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-from cmdb.data_storage.errors import DataBaseError
+from cmdb.database.errors import DataBaseError
 
 
 class DatabaseConnectionError(DataBaseError):
