+++ cmdb/database/errors/database_errors.py
