+++ cmdb/database/utils.py
