+++ cmdb/file_export/export_types.py
@@ -25,7 +25,7 @@ import zipfile
 import openpyxl
 from cmdb.utils.helpers import load_class
 
-from cmdb.data_storage.database_manager import DatabaseManagerMongo
+from cmdb.database.database_manager import DatabaseManagerMongo
 from cmdb.framework.managers.type_manager import TypeManager
 from cmdb.framework.cmdb_object_manager import CmdbObjectManager
 from cmdb.utils import json_encoding
