+++ cmdb/interface/rest_api/__init__.py
@@ -42,7 +42,7 @@ def create_rest_api(event_queue):
     system_config_reader = SystemConfigReader()
 
     # Create managers
-    from cmdb.data_storage.database_manager import DatabaseManagerMongo
+    from cmdb.database.database_manager import DatabaseManagerMongo
     app_database = DatabaseManagerMongo(
         **system_config_reader.get_all_values_from_section('Database')
     )
