+++ cmdb/media_library/media_file_manager.py
@@ -18,7 +18,7 @@ import logging
 
 from datetime import datetime
 
-from cmdb.data_storage.database_manager import DatabaseManagerMongo, DatabaseGridFS
+from cmdb.database.database_manager import DatabaseManagerMongo, DatabaseGridFS
 from gridfs.grid_file import GridOutCursor, GridOut
 
 from cmdb.media_library.media_file import MediaFile
