+++ cmdb/security/auth/providers/external_providers.py
@@ -17,7 +17,7 @@ import logging
 
 from datetime import datetime
 
-from cmdb.data_storage.database_manager import DatabaseManagerMongo
+from cmdb.database.database_manager import DatabaseManagerMongo
 from cmdb.manager.errors import ManagerGetError, ManagerInsertError
 from cmdb.security.auth.auth_errors import AuthenticationError
 from cmdb.security.auth.auth_providers import AuthenticationProvider
