+++ cmdb/user_management/managers/user_manager.py
@@ -15,7 +15,7 @@
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 from typing import Union, List
 
-from cmdb.data_storage.database_manager import DatabaseManagerMongo
+from cmdb.database.database_manager import DatabaseManagerMongo
 from cmdb.user_management.models.user import UserModel
 from .account_manager import AccountManager
 from .. import UserGroupModel
